+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-DEVEL - **0.1.x**
- * New interactive world approach [(rctoris)](https://github.com/rctoris/)
+2014-11-20 - **0.1.0**
+ * New interactive world approach with new RMS [(rctoris)](https://github.com/rctoris/)
 
 2013-06-03 - **0.0.2**
  * Adding of save and load functions [(rctoris)](https://github.com/rctoris/)